api/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-jose
pytest
